Renaming the comandline script
When running with debugpy, the name fo the package matching
the name of the cli script produced an error in identifying
the package names in the include path making a crash at the
time of loading the configuration.
