+++ pyproject.toml
@@ -41,7 +41,7 @@ dev = [
 ]
 
 [project.scripts]
-dd-license-attribution = "dd_license_attribution.cli.dd_license_attribution:app"
+dd-license-attribution = "dd_license_attribution.cli.main_cli:app"
 
 [tool.setuptools.packages.find]
 where = ["src"]
