+++ tests/unit/test_main_cli.py
@@ -7,7 +7,7 @@ import os
 
 from typer.testing import CliRunner
 
-from dd_license_attribution.cli.dd_license_attribution import app
+from dd_license_attribution.cli.main_cli import app
 
 runner = CliRunner(mix_stderr=False)
